# docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Risk Management System API",
        "version": "1.0.0",
        "description": "A comprehensive Risk Management Backend System for portfolio risk assessment, scoring, and analysis capabilities for individual investors. Designed for integration with MuleSoft.\n\n## Overview\nThis API provides:\n- **Risk Tolerance Assessment** - 15-question questionnaire processing and scoring\n- **Portfolio Risk Analysis** - VaR, Sharpe Ratio, Beta, Standard Deviation calculations\n- **Concentration Risk Detection** - Single position, sector, and top-5 holdings analysis\n- **Stress Testing** - Scenario analysis with historical and hypothetical scenarios\n- **Investment Suitability** - Portfolio vs risk profile alignment checks\n- **Recommendations** - Generated investment recommendations\n\n## Authentication\nThis is a demo API with **no authentication required**.\n\n## Base URLs\n- **Production:** https://risk-management-system-f04b786dc797.herokuapp.com/api/v1\n- **Local:** http://localhost:3002/api/v1\n\n## Data Note\nThis system uses an **in-memory datastore** that resets when the server restarts. It comes pre-loaded with 50 sample investors.",
        "contact": {
            "name": "API Support",
            "email": "support@riskmanagement.demo"
        },
        "license": {
            "name": "ISC",
            "url": "https://opensource.org/licenses/ISC"
        }
    },
    "servers": [
        {
...
        },
        {
...
    "components": {
        "schemas": {
            "Investor": {
                "type": "object",
                "properties": {
                    "investorId": {
                        "type": "string",
                        "example": "INV-001"
                    },
                    "customerId": {
                        "type": "string",
                        "example": "CUST-001"
                    },
                    "firstName": {
                        "type": "string",
                        "example": "John"
                    },
                    "lastName": {
                        "type": "string",
                        "example": "Smith"
                    },
                    "email": {
                        "type": "string",
                        "example": "john.smith@email.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "555-0101"
                    },
                    "dateOfBirth": {
                        "type": "string",
                        "format": "date",
                        "example": "1975-03-15"
                    },
                    "age": {
                        "type": "integer",
                        "example": 49
                    },
                    "employmentStatus": {
                        "type": "string",
                        "enum": [
                            "EMPLOYED",
                            "SELF_EMPLOYED",
                            "RETIRED",
                            "UNEMPLOYED"
                        ],
                        "example": "EMPLOYED"
                    },
                    "annualIncome": {
                        "type": "number",
                        "example": 150000
                    },
                    "netWorth": {
                        "type": "number",
                        "example": 850000
                    },
                    "liquidNetWorth": {
                        "type": "number",
                        "example": 450000
                    },
                    "investmentExperience": {
                        "type": "string",
                        "enum": [
                            "NOVICE",
                            "INTERMEDIATE",
                            "EXPERIENCED",
                            "EXPERT"
                        ],
                        "example": "EXPERIENCED"
                    },
                    "investmentHorizon": {
                        "type": "string",
                        "enum": [
                            "SHORT_TERM",
                            "MEDIUM_TERM",
                            "LONG_TERM"
                        ],
                        "example": "LONG_TERM"
                    }
                }
            },
...
        }
    },
    "paths": {
...
        "/investors/{investorId}/profile": {
            "get": {
                "summary": "Get investor with active risk profile",
                "description": "Returns investor details along with their current active risk profile",
                "tags": [
                    "Investors"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "investorId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "The investor ID",
                        "example": "INV-001"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Investor with profile found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "investor": {
                                                    "$ref": "#/components/schemas/Investor"
                                                },
                                                "riskProfile": {
                                                    "type": "object",
                                                    "properties": {
                                                        "profileId": {
                                                            "type": "string",
                                                            "example": "PRF-001"
                                                        },
                                                        "riskCategory": {
                                                            "type": "string",
                                                            "example": "MODERATE"
                                                        },
                                                        "compositeRiskScore": {
                                                            "type": "number",
                                                            "example": 65
                                                        },
                                                        "recommendedAllocation": {
                                                            "type": "object",
                                                            "properties": {
                                                                "equities": {
                                                                    "type": "integer"
                                                                },
                                                                "fixedIncome": {
                                                                    "type": "integer"
                                                                },
                                                                "alternatives": {
                                                                    "type": "integer"
                                                                },
                                                                "cash": {
                                                                    "type": "integer"
                                                                }
                                                            }
                                                        },
                                                        "maxDrawdownTolerance": {
                                                            "type": "number",
                                                            "example": 25
                                                        },
                                                        "maxVolatilityTolerance": {
                                                            "type": "number",
                                                            "example": 18
                                                        },
                                                        "lastAssessmentDate": {
                                                            "type": "string",
                                                            "format": "date-time"
                                                        }
                                                    }
                                                }
                                            }
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Investor not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/profiles": {
            "get": {
                "summary": "List all risk profiles",
                "description": "Returns all risk profiles across all investors",
                "tags": [
                    "Profiles"
                ],
                "responses": {
                    "200": {
                        "description": "List of all risk profiles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "profiles": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/RiskProfile"
                                                    }
                                                },
                                                "count": {
                                                    "type": "integer",
                                                    "example": 50
                                                }
                                            }
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create profile (admin override)",
                "description": "Manually creates a risk profile without going through assessment",
                "tags": [
                    "Profiles"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "investorId",
                                    "riskCategory"
                                ],
                                "properties": {
                                    "investorId": {
                                        "type": "string",
                                        "example": "INV-001"
                                    },
                                    "riskCategory": {
                                        "type": "string",
                                        "enum": [
                                            "CONSERVATIVE",
                                            "MODERATELY_CONSERVATIVE",
                                            "MODERATE",
                                            "MODERATELY_AGGRESSIVE",
                                            "AGGRESSIVE"
                                        ]
                                    },
                                    "compositeRiskScore": {
                                        "type": "number"
                                    },
                                    "recommendedAllocation": {
                                        "type": "object"
                                    },
                                    "maxDrawdownTolerance": {
                                        "type": "number"
                                    },
                                    "maxVolatilityTolerance": {
                                        "type": "number"
                                    },
                                    "maxConcentrationLimit": {
                                        "type": "number"
                                    },
                                    "isActive": {
                                        "type": "boolean",
                                        "default": true
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Profile created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/RiskProfile"
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/profiles/{profileId}": {
            "get": {
                "summary": "Get profile by ID",
                "description": "Returns a single risk profile by its ID",
                "tags": [
                    "Profiles"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "profileId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "The profile ID",
                        "example": "PRF-001"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Profile found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/RiskProfile"
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Risk profile not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "summary": "Update profile",
                "description": "Updates an existing risk profile",
                "tags": [
                    "Profiles"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "profileId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "The profile ID",
                        "example": "PRF-001"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "riskCategory": {
                                        "type": "string",
                                        "enum": [
                                            "CONSERVATIVE",
                                            "MODERATELY_CONSERVATIVE",
                                            "MODERATE",
                                            "MODERATELY_AGGRESSIVE",
                                            "AGGRESSIVE"
                                        ]
                                    },
                                    "compositeRiskScore": {
                                        "type": "number"
                                    },
                                    "recommendedAllocation": {
                                        "type": "object"
                                    },
                                    "maxDrawdownTolerance": {
                                        "type": "number"
                                    },
                                    "maxVolatilityTolerance": {
                                        "type": "number"
                                    },
                                    "isActive": {
                                        "type": "boolean"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/RiskProfile"
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Risk profile not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}